TASK: Cleanup DisabledAutoparagraphMode
