+++ packages/neos-ui-ckeditor5-bindings/src/manifest.config.js
@@ -2,7 +2,7 @@ import CkEditorConfigRegistry from './registry/CkEditorConfigRegistry';
 import {$add, $get, $or} from 'plow-js';
 import {stripTags} from '@neos-project/utils-helpers';
 
-import InlineMode from './plugins/inlineMode';
+import DisabledAutoparagraphMode from './plugins/disabledAutoparagraphMode';
 import Sub from './plugins/sub';
 import Sup from './plugins/sup';
 import LinkTargetBlank from './plugins/linkTargetBlank';
@@ -33,7 +33,7 @@ const addPlugin = (Plugin, isEnabled) => (ckEditorConfiguration, options) => {
 
 // If the editable is a span or a heading, we automatically disable paragraphs and enable the soft break mode
 // Also possible to force this behavior with `autoparagraph: false`
-const disableParagraph = (editorOptions, {propertyDomNode}) =>
+const disableAutoparagraph = (editorOptions, {propertyDomNode}) =>
     $get('autoparagraph', editorOptions) === false ||
     propertyDomNode.tagName === 'SPAN' ||
     propertyDomNode.tagName === 'H1' ||
@@ -100,7 +100,7 @@ export default ckEditorRegistry => {
     //
     config.set('essentials', addPlugin(Essentials));
     config.set('paragraph', addPlugin(Paragraph));
-    config.set('inlineMode', addPlugin(InlineMode, disableParagraph));
+    config.set('disabledAutoparagraphMode', addPlugin(DisabledAutoparagraphMode, disableAutoparagraph));
     config.set('sub', addPlugin(Sub, $get('formatting.sub')));
     config.set('sup', addPlugin(Sup, $get('formatting.sup')));
     config.set('bold', addPlugin(Bold, $get('formatting.strong')));
